backend/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
PyJWT==1.7.1
beautifulsoup4==4.7.1
flask-mongoengine==0.9.5
gunicorn==19.9.0
python-cas==1.4.0
requests==2.22.0
mongomock==3.21.0
pytest==6.1.2
coverage==5.3
